# docs/swagger.yaml
basePath: /
definitions:
  api.ActressList:
    properties:
      action:
        example: a.CreatedAt
        type: string
      actress:
        example: ""
        type: string
      page:
        type: integer
      size:
        type: integer
      sort:
        example: desc
        type: string
    type: object
  api.ChangePassword:
    properties:
      confirm_password:
        type: string
      new_password:
        type: string
      old_password:
        type: string
    required:
    - confirm_password
    - new_password
    - old_password
    type: object
  api.DanmuSave:
    properties:
      border:
        type: boolean
      color:
        type: string
      mode:
        type: integer
      style:
        type: string
      text:
        type: string
      time:
        type: number
      video_id:
        type: integer
    required:
    - color
    - text
    - video_id
    type: object
  api.Fail:
    properties:
      code:
        example: 0
        type: integer
      message:
        example: status bad request
        type: string
      status:
        example: false
        type: boolean
    type: object
  api.ForgotPassword:
    properties:
      confirm_password:
        type: string
      password:
        type: string
      reset_password_token:
        type: string
    required:
    - confirm_password
    - password
    - reset_password_token
    type: object
  api.Login:
    properties:
      email:
        example: wxw9868@163.com
        type: string
      password:
        example: "123456"
        type: string
    required:
    - email
    - password
    type: object
  api.Message:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      status:
        type: boolean
    type: object
  api.NotFound:
    properties:
      code:
        example: 404
        type: integer
      message:
        example: status not found
        type: string
      status:
        example: false
        type: boolean
    type: object
  api.Register:
    properties:
      email:
        type: string
      password:
        type: string
      repeat_password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - repeat_password
    - username
    type: object
  api.ServerError:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: status internal server error
        type: string
      status:
        example: false
        type: boolean
    type: object
  api.Success:
    properties:
      code:
        example: 1
        type: integer
      data: {}
      message:
        example: status ok
        type: string
      status:
        example: true
        type: boolean
    type: object
  api.UserUpdate:
    properties:
      address:
        type: string
      designation:
        type: string
      email:
        type: string
      mobile:
        type: string
      nickname:
        type: string
      note:
        type: string
      username:
        type: string
    required:
    - email
    - mobile
    - nickname
    - username
    type: object
  api.Video:
    properties:
      action:
        example: v.CreatedAt
        type: string
      actress_id:
        type: integer
      page:
        type: integer
      size:
        type: integer
      sort:
        example: desc
        type: string
    type: object
  api.VideoCollect:
    properties:
      collect:
        enum:
        - 1
        - -1
        type: integer
      video_id:
        type: integer
    required:
    - collect
    - video_id
    type: object
host: 192.168.0.9:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a video server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Video API
  version: "1.0"
paths:
  /actress/list:
    post:
      consumes:
      - application/json
      description: 演员列表
      parameters:
      - description: 演员列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.ActressList'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Fail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ServerError'
      summary: 演员列表
      tags:
      - actress
  /danmu/list/{video_id}:
    get:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: Video ID
        in: path
        name: video_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Fail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ServerError'
      summary: 视频弹幕列表
      tags:
      - video
  /danmu/save:
    post:
      consumes:
      - application/json
      description: 视频弹幕
      parameters:
      - description: 视频弹幕
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.DanmuSave'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Fail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ServerError'
      summary: 视频弹幕
      tags:
      - danmu
...
  /stock/liquidation:
    get:
      consumes:
      - application/json
      description: 用于已清仓股票
      parameters:
      - description: 页码
        in: query
        name: page
        type: string
      - description: 每页条数
        in: query
        name: page_size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Message'
      summary: 已清仓股票
      tags:
      - stock
  /stock/tradingRecords:
    get:
      consumes:
      - application/json
      description: 历史成交数据
      parameters:
      - description: 页码
        in: query
        name: page
        type: string
      - description: 每页条数
        in: query
        name: page_size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Message'
      summary: 历史成交
      tags:
      - stock
  /user/avatar:
    post:
      consumes:
      - multipart/form-data
      description: Upload file
      parameters:
      - description: account avatar
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Fail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ServerError'
      summary: Upload account avatar
      tags:
      - user
  /user/browse:
    get:
      consumes:
      - application/json
      description: 用户浏览记录
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Success'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ServerError'
      summary: 用户浏览记录
      tags:
      - user
  /user/changePassword:
    post:
      consumes:
      - application/json
      description: 用户修改密码
      parameters:
      - description: 用户修改密码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.ChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Fail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ServerError'
      summary: 用户修改密码
      tags:
      - user
  /user/collect:
    get:
      consumes:
      - application/json
      description: 用户收藏记录
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Success'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ServerError'
      summary: 用户收藏记录
      tags:
      - user
  /user/forgotPassword:
    post:
      consumes:
      - application/json
      description: 用户忘记密码
      parameters:
      - description: 用户忘记密码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.ForgotPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Fail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ServerError'
      summary: 用户忘记密码
      tags:
      - user
  /user/info:
    get:
      consumes:
      - application/json
      description: 用户信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Success'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ServerError'
      summary: 用户信息
      tags:
      - user
  /user/login:
    post:
      consumes:
      - application/json
      description: 用户登录
      parameters:
      - description: 用户登录
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Fail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ServerError'
      summary: 用户登录
      tags:
      - user
  /user/register:
    post:
      consumes:
      - application/json
      description: 用户注册
      parameters:
      - description: 用户注册
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.Register'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Fail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ServerError'
      summary: 用户注册
      tags:
      - user
  /user/save:
    post:
      consumes:
      - application/json
      description: 用户修改信息
      parameters:
      - description: 用户修改信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.UserUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Fail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ServerError'
      summary: 用户修改信息
      tags:
      - user
  /user/session:
    get:
      consumes:
      - application/json
      description: 用户信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Success'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.NotFound'
      summary: 用户信息
      tags:
      - user
  /video/browse/{video_id}:
    get:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: Video ID
        in: path
        name: video_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Fail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ServerError'
      summary: 视频浏览记录
      tags:
      - video
  /video/collect:
    post:
      consumes:
      - application/json
      description: 视频收藏
      parameters:
      - description: 视频收藏
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.VideoCollect'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Fail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ServerError'
      summary: 视频收藏
      tags:
      - video
  /video/import:
    get:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: dir
        in: query
        name: dir
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Fail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ServerError'
      summary: 视频导入
      tags:
      - video
  /video/list:
    post:
      consumes:
      - application/json
      description: 视频列表
      parameters:
      - description: 视频列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.Video'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Fail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ServerError'
      summary: 视频列表
      tags:
      - video
  /video/play/{id}:
    get:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Fail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ServerError'
      summary: 视频信息
      tags:
      - video
  /video/search:
    get:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: 关键词
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Fail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ServerError'
      summary: 视频搜索
      tags:
      - video
swagger: "2.0"